add listview with RealmBaseAdapter

+++ Android_TodoList/app/src/main/java/com/example/todolist/Todo.kt
@@ -4,7 +4,7 @@ import io.realm.RealmObject
 import io.realm.annotations.PrimaryKey
 
 open class Todo (
-    @PrimaryKey var id: Int = 0,
+    @PrimaryKey var id: Long = 0,
     var title: String = "",
     var date: Long = 0
 ): RealmObject()
